+++ pom.xml
@@ -68,6 +68,11 @@
 		  <artifactId>mysql-connector-java</artifactId>
 		  <scope>runtime</scope>
 		</dependency>
+		<!-- SpringSecurity -->
+		<dependency>
+        	<groupId>org.springframework.boot</groupId>
+        	<artifactId>spring-boot-starter-security</artifactId>
+    	</dependency>
 		<!-- jQuery -->
 		<dependency>
 		  <groupId>org.webjars</groupId>
