+++ src/main/resources/templates/error.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+  <meta charset="UTF-8">
+  <title>/error</title>
+</head>
+<body>
+<div th:text="'timestamp: ' + ${timestamp}"></div>
+<div th:text="'status: ' + ${status}"></div>
+<div th:text="'error: ' + ${error}"></div>
+<div th:text="'exception: ' + ${exception}"></div>
+<div th:text="'message: ' + ${message}"></div>
+<div th:text="'errors: ' + ${errors}"></div>
+<div th:text="'trace: ' + ${trace}"></div>
+<div th:text="'path: ' + ${path}"></div>
+</body>
+</html>
