+++ src/main/resources/templates/login.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+	<meta charset="UTF-8">
+	<link th:href="@{/webjars/bootstrap/3.3.7-1/css/bootstrap.min.css}" rel="stylesheet" />
+<link th:href="@{/common/css/common.css}" rel="stylesheet" />
+<script th:src="@{/webjars/jquery/1.11.1/jquery.min.js}"></script>
+<script th:src="@{/webjars/bootstrap/3.3.7-1/js/bootstrap.min.js}"></script>
+    <title>ログインページ</title>
+</head>
+<body class="text-center">
+	<h1>Login</h1>
+    <form action="/login" method="post">
+    <p th:if="${session[SPRING_SECURITY_LAST_EXCEPTION]}!=NULL"
+    	th:text="${session[SPRING_SECURITY_LAST_EXCEPTION].message}"
+    		class="text-danger">
+    		ログインエラーメッセージ
+    		</p>
+        <div>
+            <label>ユーザーID</label>
+            <input type="text" id="user_id" name="user_id"/>
+        </div>
+        <div>
+            <label for="password">パスワード</label>
+            <input type="password" id="password" name="password"/>
+        </div>
+        <div>
+            <button class="btn btn-primary" type="submit">ログイン</button>
+        </div>
+    </form>
+    <br />
+    <a th:href="@{'signup'}">ユーザー新規登録はこちら</a>
+</body>
+</html>
